Adapt code to new workflow
The new workflow performs a local check for obsolete person and organisation document and does not rely anymore on advanced search queries in Elasticsearch itself. For that matter a first cycle gets all contributor identifiers in type `bibliographicResource`. Next all identifiers for `organisation` and `person` documents are fetched. Based on a comparison between these datasets, a bulk delete on type `person` and `organisation` is performed eventually.

Signed-off-by: Sebastian Schüpbach <sebastian.schuepbach@unibas.ch>